set init values
